Added { signed: true } option to res.cookie()
