+++ lib/response.js
@@ -441,29 +441,12 @@ res.clearCookie = function(name, options){
 };
 
 /**
- * Set a signed cookie with the given `name` and `val`.
- * See `res.cookie()` for details.
- *
- * @param {String} name
- * @param {String|Object} val
- * @param {Object} options
- * @api public
- */
-
-res.signedCookie = function(name, val, options){
-  var secret = this.req.secret;
-  if (!secret) throw new Error('connect.cookieParser("secret") required for signed cookies');
-  if ('object' == typeof val) val = 'j:' + JSON.stringify(val);
-  val = utils.sign(val, secret);
-  return this.cookie(name, val, options);
-};
-
-/**
  * Set cookie `name` to `val`, with the given `options`.
  *
  * Options:
  *
  *    - `maxAge`   max-age in milliseconds, converted to `expires`
+ *    - `signed`   sign the cookie
  *    - `path`     defaults to "/"
  *
  * Examples:
@@ -482,7 +465,11 @@ res.signedCookie = function(name, val, options){
 
 res.cookie = function(name, val, options){
   options = options || {};
+  var secret = this.req.secret;
+  var signed = options.signed;
+  if (signed && !secret) throw new Error('connect.cookieParser("secret") required for signed cookies');
   if ('object' == typeof val) val = 'j:' + JSON.stringify(val);
+  if (signed) val = utils.sign(val, secret);
   if ('maxAge' in options) options.expires = new Date(Date.now() + options.maxAge);
   if (null == options.path) options.path = '/';
   var cookie = utils.serializeCookie(name, val, options);
